Add files via upload
Updated code I uploaded yesterday. Fully imported and cleaned data to have a 3D dataframe (country and year on X, measure name on Y) that's ready to start analyzing and looking for outliers etc.. I've attacked a pkl of the ready-to-go data as well as the csv of countries to use that's found on Blackboard. Finally, I copied instructions from Blackboard into the script to try to get all the information we would need in one place, and made a few small functions for getting specific slices of the data.